fix the issue of deleting the xml file when error occurs.

+++ src/main/java/com/redpill/api/DatabaseProxy.java
@@ -2,6 +2,7 @@ package com.redpill.api;
 
 import com.redpill.entity.DatabaseEntity;
 import com.redpill.server.AnaTask;
+import com.redpill.tool.FileTool;
 import com.redpill.tool.LogTool;
 import com.redpill.tool.MuleConfig;
 import com.redpill.tool.RedisUtils;
@@ -14,8 +15,10 @@ import org.mule.api.config.ConfigurationException;
 import org.mule.api.lifecycle.InitialisationException;
 import org.mule.config.spring.SpringXmlConfigurationBuilder;
 import org.mule.context.DefaultMuleContextFactory;
+import org.omg.CORBA.PUBLIC_MEMBER;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
@@ -23,6 +26,7 @@ import java.util.List;
 import static com.redpill.CFManager.saveDocument;
 
 public class DatabaseProxy implements MuleTask{
+    public static final String TYPE = "T13";
     private static final String xmlTemplatePath = "database.xml";
     private static final DefaultMuleContextFactory defaultMuleContextFactory = new DefaultMuleContextFactory();
     String xmlPre = "HDB-";
@@ -32,12 +36,18 @@ public class DatabaseProxy implements MuleTask{
     public MuleContext muleContext;
 
     private DatabaseEntity databaseEntity;
+    private String serverId;
 
     public DatabaseProxy(DatabaseEntity databaseEntity) {
         this.databaseEntity = databaseEntity;
-        this.xmlName = xmlPre + databaseEntity.getTask_id() + ".xml";
-    }
 
+        this.serverId = getServerId();
+        this.xmlName = xmlPre + serverId + ".xml";
+    }
+    @Override
+    public String getType() {
+        return TYPE;
+    }
 //TEST
 public static void main(String[] args) throws InterruptedException {
     try {
@@ -83,12 +93,13 @@ public static void main(String[] args) throws InterruptedException {
             configBuilder = new SpringXmlConfigurationBuilder(xmlPath + xmlName);
             muleContext = defaultMuleContextFactory.createMuleContext(configBuilder);
             muleContext.start();
-            String id = muleContext.getConfiguration().getId();
-            String taskInfo = RedisUtils.redisPool.jedis(jedis -> {
-                jedis.hset(MuleConfig.muleMonitor + databaseEntity.getTask_id(), MuleConfig.hostIp, id);
+            String muleId = muleContext.getConfiguration().getId();
+            RedisUtils.redisPool.jedis(jedis -> {
+                jedis.hset(MuleConfig.muleMonitor, getTaskId(), serverId);
+                jedis.hset(MuleConfig.muleMonitor + serverId, MuleConfig.hostIp, muleId);
                 return null;
             });
-            AnaTask.addToTaskMap(databaseEntity.getTask_id(), this);
+            AnaTask.addToTaskMap(getServerId(), this);
             return true;
         } catch (Exception e) {
             LogTool.logInfo(1, databaseEntity.getTask_id() + " start error.");
@@ -104,7 +115,7 @@ public static void main(String[] args) throws InterruptedException {
             try {
                 muleContext.stop();
                 RedisUtils.redisPool.jedis(jedis -> {
-                    jedis.del(MuleConfig.muleMonitor + databaseEntity.getTask_id());
+                    jedis.hdel(MuleConfig.muleMonitor, getTaskId());
                     return null;
                 });
             } catch (MuleException e) {
@@ -117,13 +128,67 @@ public static void main(String[] args) throws InterruptedException {
     }
 
     @Override
-    public void removeTask() {
-        if (muleContext != null){
-            this.closeTask();
-        }
+    public boolean removeTask() {
         File file = new File(xmlPath + xmlName);
-        file.delete();
-        LogTool.logInfo(2, "rm task " + databaseEntity.getTask_id());
+        SAXBuilder saxBuilder = new SAXBuilder();
+        InputStream is = ClassLoader.getSystemResourceAsStream(xmlPath + xmlName);
+        Document document = null;
+        try {
+            document = saxBuilder.build(is);
+        } catch (JDOMException e) {
+            e.printStackTrace();
+            return false;
+        } catch (IOException e) {
+            e.printStackTrace();
+            return false;
+        }
+        Element rootElement = document.getRootElement();
+        List<Element> children = rootElement.getChildren();
+
+        for (Element child : children) {
+            if ((child.getName().equalsIgnoreCase("flow") || (child.getName().equalsIgnoreCase("generic-config ")))
+                    && child.getAttribute("name").getValue().equalsIgnoreCase(getTaskId())) {
+                rootElement.removeChild(child.getName(), child.getNamespace());
+            }
+        }
+        boolean haveFlows = false;
+        for (Element child : children) {
+            if (child.getName().equalsIgnoreCase("flow")){
+                haveFlows =  true;
+            }
+        }
+        if (haveFlows){
+            try {
+                saveDocument(document, file);
+                LogTool.logInfo(2, "error, delete task " + getTaskId() + ", re-save file.");
+            } catch (IOException e) {
+                e.printStackTrace();
+                return false;
+            }
+        }else {
+            LogTool.logInfo(2, "error, delete task " + getTaskId() + ", delete file.");
+            file.delete();
+            return false;
+        }
+
+        LogTool.logInfo(2, "rm task " + getTaskId());
+
+        return false;
+    }
+
+    @Override
+    public String getTaskId() {
+        return databaseEntity.getTask_id();
+    }
+
+    @Override
+    public String getServerId() {
+        return TYPE + "-" + databaseEntity.getConfig().getHttp_port();
+    }
+
+    @Override
+    public Integer getPort() {
+        return databaseEntity.getConfig().getHttp_port();
     }
 
     @Override
@@ -142,49 +207,73 @@ public static void main(String[] args) throws InterruptedException {
         }
     }
     private String databaseProxy() throws IOException, JDOMException {
+
+        File file = new File(xmlPath + xmlName);
+        boolean exist = true;
+        if (!file.exists()){
+            file.createNewFile();
+            exist = false;
+        }else {
+            String backupFileName = xmlPath + xmlName + ".backup";
+            File fileBackup = new File(backupFileName);
+            if (!fileBackup.exists()) {
+                fileBackup.createNewFile();
+                FileTool.copyFile(xmlPath + xmlName, backupFileName);
+            }
+        }
+
         SAXBuilder saxBuilder = new SAXBuilder();
-        InputStream is = ClassLoader.getSystemResourceAsStream(xmlTemplatePath);
+        InputStream is = null;
+        if(exist){
+            is = new FileInputStream(xmlPath + xmlName);
+        }else{
+            is = ClassLoader.getSystemResourceAsStream(xmlTemplatePath);
+        }
         Document document = saxBuilder.build(is);
         Element rootElement = document.getRootElement();
         List<Element> children = rootElement.getChildren();
-        String httpConfName = databaseEntity.getTask_id() + "-listener-conf";
-        String databaseConfName = databaseEntity.getTask_id() + "-db-conf";
         Namespace httpNs = Namespace.getNamespace("http", "http://www.mulesoft.org/schema/mule/http");
         Namespace dbNs = Namespace.getNamespace("db", "http://www.mulesoft.org/schema/mule/db");
         Namespace defaultNs = Namespace.getNamespace("", "http://www.mulesoft.org/schema/mule/core");
         Namespace docNs = Namespace.getNamespace("doc", "http://www.mulesoft.org/schema/mule/documentation");
         Namespace jsonNs = Namespace.getNamespace("json", "http://www.mulesoft.org/schema/mule/json");
 
-        for (Element child : children) {
-            if (child.getName().equals("listener-config")){
-                child.getAttribute("name").setValue(httpConfName);
-                child.getAttribute("host").setValue(databaseEntity.getConfig().getHttp_host());
-                child.getAttribute("port").setValue(String.valueOf(databaseEntity.getConfig().getHttp_port()));
-                child.getAttribute("name", docNs).setValue(httpConfName);
-            }else if (child.getName().equals("generic-config")){
-                if (databaseEntity.getConfig().getData_type().equalsIgnoreCase("oracle")){
-                    child.setName("oracle-config");
-                    child.getAttribute("name").setValue(databaseConfName);
-                    child.getAttribute("name", docNs).setValue(databaseConfName);
-                    child.setAttribute("host", databaseEntity.getConfig().getIp_address());
-                    child.setAttribute("port", String.valueOf(databaseEntity.getConfig().getPort()));
-                    child.setAttribute("instance", databaseEntity.getConfig().getDatabase_ins());
-                    child.setAttribute("user", databaseEntity.getConfig().getUser_name());
-                    child.setAttribute("password", databaseEntity.getConfig().getPassword());
-                }else {
-                    child.getAttribute("name").setValue(databaseConfName);
-                    child.getAttribute("url").setValue(databaseEntity.getConfig().getDatabase_url());
-                    child.getAttribute("driverClassName").setValue(databaseEntity.getConfig().getDriverClassName());
-                    child.getAttribute("name", docNs).setValue(databaseConfName);
-                }
-            }
+        String listenerConfName = getServerId() + "-in";
+        String genericConfig = getTaskId();
+        if (!exist){
+            // 创建listener config
+            Element listenerConf = new Element("listener-config", httpNs);
+            listenerConf.setAttribute("name", listenerConfName);
+            listenerConf.setAttribute("host", "0.0.0.0");
+            listenerConf.setAttribute("port", String.valueOf(getPort()));
+            listenerConf.setAttribute("name", listenerConfName+" configuration", docNs);
+            rootElement.addContent(0, listenerConf);
         }
+
+        Element elementGenericConfig = new Element("generic-config", dbNs);
+        if (databaseEntity.getConfig().getData_type().equalsIgnoreCase("oracle")){
+            elementGenericConfig.setName("oracle-config");
+            elementGenericConfig.setAttribute("name", genericConfig);
+            elementGenericConfig.setAttribute("name", genericConfig, docNs);
+            elementGenericConfig.setAttribute("host", databaseEntity.getConfig().getIp_address());
+            elementGenericConfig.setAttribute("port", String.valueOf(databaseEntity.getConfig().getPort()));
+            elementGenericConfig.setAttribute("instance", databaseEntity.getConfig().getDatabase_ins());
+            elementGenericConfig.setAttribute("user", databaseEntity.getConfig().getUser_name());
+            elementGenericConfig.setAttribute("password", databaseEntity.getConfig().getPassword());
+        }else {
+            elementGenericConfig.getAttribute("name").setValue(genericConfig);
+            elementGenericConfig.getAttribute("url").setValue(databaseEntity.getConfig().getDatabase_url());
+            elementGenericConfig.getAttribute("driverClassName").setValue(databaseEntity.getConfig().getDriverClassName());
+            elementGenericConfig.getAttribute("name", docNs).setValue(genericConfig);
+        }
+        rootElement.addContent(elementGenericConfig);
+
         //add flow
         for (int i = 0; i < databaseEntity.getFlows().size(); i++) {
             Element wsd = new Element( "flow", defaultNs);
-            wsd.setAttribute("name", "flow-" + i);
+            wsd.setAttribute("name", getTaskId());
             Element listener = new Element("listener", httpNs);
-            listener.setAttribute("config-ref", httpConfName);
+            listener.setAttribute("config-ref", listenerConfName);
             listener.setAttribute("path", databaseEntity.getFlows().get(i).getTEsb().getPath());
             if (!(databaseEntity.getFlows().get(i).getTEsb().getMethod()).contains("*")){
                 listener.setAttribute("allowedMethods", databaseEntity.getFlows().get(i).getTEsb().getMethod());
@@ -208,7 +297,7 @@ public static void main(String[] args) throws InterruptedException {
             }else {
                 throw new IllegalArgumentException("no such sb operate '" + databaseEntity.getFlows().get(i).getSEsb().getOperate() + "'.");
             }
-            dbOpe.setAttribute("config-ref", databaseConfName);
+            dbOpe.setAttribute("config-ref", genericConfig);
             dbOpe.setAttribute("name", "database", docNs);
             Element dbSql = new Element("dynamic-query", dbNs);
 //            dbSql.addContent("<![CDATA["+ databaseEntity.getFlows().get(i).getSEsb().getSql()+"]]>");
@@ -246,10 +335,6 @@ public static void main(String[] args) throws InterruptedException {
             wsd.addContent(exception);
 
             rootElement.addContent(wsd);
-        }
-        File file = new File(xmlPath + xmlName);
-        if (!file.exists()){
-            file.createNewFile();
         }
         saveDocument(document, file);
         return xmlPath + xmlName;
